Fixed return type for Groovy countBy.

+++ groovy-module/src/main/groovy/jvm/language/compare/groovy/cards/GroovyDeckOfCards.groovy
@@ -48,11 +48,11 @@ class GroovyDeckOfCards {
         this.cardsBySuit.get(Suit.CLUBS)
     }
 
-    Map<Suit, Long> countsBySuit() {
+    Map<Suit, Integer> countsBySuit() {
         this.cards.countBy { it.getSuit() }
     }
 
-    Map<Rank, Long> countsByRank() {
+    Map<Rank, Integer> countsByRank() {
         this.cards.countBy { it.getRank() }
     }
